requirements.txt
python-dotenv>=1.0.0
huggingface-hub>=0.19.0
